refactor: move deepseek/opencode model parsers to backend packages
- Create deepseek/discovery.go: move ParseDeepSeekModels as unexported
  parseDeepSeekModels, add DiscoverDeepSeekModels() that tries 'codewhale
  models' then falls back to 'deepseek models', registers via
  model.RegisterDiscoverModelsFunc in init()
- Create opencode/discovery.go: move ParseOpenCodeModels as unexported
  parseOpenCodeModels, add DiscoverOpenCodeModels() that runs
  'opencode models', registers via model.RegisterDiscoverModelsFunc
- Move parser tests to respective backend packages
- Remove ParseDeepSeekModels, ParseOpenCodeModels, regex vars from
  discovery.go; remove 'regexp' import
- Remove ListModelsCmd+ParseModels from opencode and deepseek BackendSpec
- Remove ListModelsCmd+ParseModels code path from discoverModels()
- Simplify CanDiscoverModels() to only check registry
- Update tests to use registry-based discovery pattern
